% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_sys}
\alias{app_sys}
\title{Locate packaged application files}
\usage{
app_sys(...)
}
\arguments{
\item{...}{Components passed to \code{\link[=system.file]{system.file()}}.}
}
\value{
The resolved path within the installed package.
}
\description{
Locate packaged application files
}
\keyword{internal}
