% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Application user interface}
\usage{
app_ui(request)
}
\arguments{
\item{request}{Internal Shiny request object.}
}
\value{
Shiny UI.
}
\description{
Provides the top-level navigation, including the county map view.
}
\keyword{internal}
