% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{bird_db_path}
\alias{bird_db_path}
\title{Default path for the personal bird database}
\usage{
bird_db_path()
}
\value{
A normalized path beneath the user's platform-specific application
data directory. The directory is created if necessary.
}
\description{
Default path for the personal bird database
}
