% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{get_golem_config}
\alias{get_golem_config}
\title{Access application configuration}
\usage{
get_golem_config(
  value,
  config = Sys.getenv("R_CONFIG_ACTIVE", "default"),
  use_parent = TRUE
)
}
\arguments{
\item{value}{Name of the configuration value.}

\item{config}{Active configuration profile.}

\item{use_parent}{Logical; passed to \code{golem::get_golem_config()}.}
}
\value{
A configuration value.
}
\description{
Access application configuration
}
\keyword{internal}
