% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map.R
\name{mod_map_ui}
\alias{mod_map_ui}
\title{Create the county map module UI}
\usage{
mod_map_ui(id)
}
\arguments{
\item{id}{Shiny module namespace identifier.}
}
\value{
A Shiny UI fragment containing the county map output.
}
\description{
Create the county map module UI
}
\keyword{internal}
